requirements.txt
asgiref==3.5.2
backcall==0.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
decorator==5.1.1
Django==3.2.13
django-proxy==1.2.1
djangorestframework==3.13.1
idna==3.3
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
libcst==0.4.5
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pydot==1.4.2
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.2.1.post0
typing-inspect==0.7.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
